Updated FMOD Listener in Unity, tested some sounds
Mainly attached the player object to the FMOD Listener (Main Camera), so Unity can handle distance attenuation. I tested it with the Leviathan, and it easily attenuates both the volume and a low-pass filter I threw on it.

I also updated the door sounds, but haven't tested them in Unity yet